+++ imgui/imgui-core/src/main/java/imgui/ImGuiIO.java
@@ -134,7 +134,7 @@ public class ImGuiIO extends IDLBase {
     */
     private static native boolean containsIniFilenameNATIVE(long addr);
 
-    public void SetFontTexID(int id) {
+    public void SetFontTexID(long id) {
         SetFontTexIDNATIVE(native_address, id);
     }
 
@@ -143,7 +143,7 @@ public class ImGuiIO extends IDLBase {
               ImGuiIO* io = (ImGuiIO*)addr;
               io->Fonts->TexID = (ImTextureID)id;
     */
-    private static native void SetFontTexIDNATIVE(long addr, int id);
+    private static native void SetFontTexIDNATIVE(long addr, long id);
 
     public void UpdateKeyTyped(int c) {
         updateKeyTyped(native_address, c);
